refactor(event): consolidate event handling and remove unused event classes
